requirements.txt
keyboard~=0.13.5
Pillow~=10.0.0
pyinstaller~=5.13.0